trdb_api/pyproject.toml
[tool.poetry]
name = "trdb-api"
version = "0.1.0"
description = "Technical Resource Database API"
authors = ["Ayushi Agrawal <ayushiag666@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "trdb_api"}]

[tool.poetry.dependencies]
python = "^3.10"
django = ">=4.2.0,<4.3.0"
djangorestframework = "^3.14.0"
django-filter = "^23.2"
django-guardian = "^2.4.0"
drf-spectacular = "^0.26.2"


[tool.poetry.group.dev.dependencies]
tox = "3.28.0"
django-extensions = "^3.2.3"
werkzeug = "^2.3.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
profile = "black"

[tool.black]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88
extend-exclude = '''
^.*\b(migrations)\b.*$
'''


[tool.bandit]
exclude_dirs = [
    "./.eggs",
    "./.git",
    "./.tox",
    "./build",
    "./coverage",
    "./dist",
    "./venv",
    "./__pycache__",
    "*tests*",
]


[tool.mypy]
exclude = '''
^.*\b(migrations)\b.*$
'''
ignore_missing_imports = true
